+++ packages/server/src/models/classes/notifications/email-envelope.js
@@ -4,6 +4,7 @@ const Mustache = require('mustache');
 
 class EmailEnvelope {
   /**
+   * @todo: use one file of two (notification, template)
    * @param {object} options
    * @param {string} options.subject
    * @param {string} options.body
