+++ packages/server/src/models/classes/templates/index.js
@@ -1,7 +1,5 @@
 const { EmailEnvelopeTemplate } = require('./email-envelope-template');
-const { FeedItemTemplate } = require('./feed-item-template');
 
 module.exports = {
   EmailEnvelopeTemplate,
-  FeedItemTemplate,
 };
